refactor: apply updated path aliases throughout codebase
