+++ app/api/contact/route.ts
@@ -1,17 +1,11 @@
-import { successResponse, errorResponse } from '~/lib/utils/apiResponse';
-import {
-  validateContactData,
-  type ContactFormData,
-} from '~/lib/utils/validateContactData';
-import { validateRequestData } from '~/lib/utils/validateRequestData';
+import { successResponse, errorResponse } from '~/utils/apiResponse';
+import { validateContactData, type ContactFormData } from '~/utils/validateContactData';
+import { validateRequestData } from '~/utils/validateRequestData';
 
 export async function POST(request: Request) {
   const data = await request.json();
 
-  const validationResult = validateRequestData<ContactFormData>(
-    data,
-    validateContactData,
-  );
+  const validationResult = validateRequestData<ContactFormData>(data, validateContactData);
 
   if (!validationResult.valid) {
     return errorResponse(validationResult.errors);
