+++ middleware.ts
@@ -1,5 +1,5 @@
 import createMiddleware from 'next-intl/middleware';
-import {routing} from './i18n/routing';
+import { routing } from '~/i18n/routing';
 
 export default createMiddleware(routing);
