+++ tsconfig.json
@@ -23,8 +23,8 @@
       "~/*": ["app/*"],
       "~/utils/*": ["app/lib/utils/*"],
       "~/design-system-components/*": ["app/shared/components/design-system/all-components/*"],
-      "~/components/*": ["app/shared/components"],
-      "~/hooks/*": ["app/shared/hooks"],
+      "~/components/*": ["app/shared/components/*"],
+      "~/hooks/*": ["app/shared/hooks/*"],
       "~/i18n/*": ["i18n/*"]
     },
     "forceConsistentCasingInFileNames": true, // Enforce consistent casing in file names
